- refactored to remove unused colors - removed portal URL and client ID from strings array - refactored module name - refactored application id - removed test depedency
