--- author-map/src/test/java/com/esri/arcgisruntime/sample/authoramap/ExampleUnitTest.java
@@ -1,15 +1,0 @@
-package com.esri.arcgisruntime.sample.authoramap;
-
-import org.junit.Test;
-
-import static org.junit.Assert.assertEquals;
-
-/**
- * To work on unit tests, switch the Test Artifact in the Build Variants view.
- */
-public class ExampleUnitTest {
-    @Test
-    public void addition_isCorrect() throws Exception {
-        assertEquals(4, 2 + 2);
-    }
-}
