+++ authoring-map/.gitignore
