+++ authoring-map/build.gradle
@@ -5,7 +5,7 @@ android {
     buildToolsVersion rootProject.ext.buildToolsVersion
 
     defaultConfig {
-        applicationId "com.example.authoramap"
+        applicationId "com.esri.arcgisruntime.sample.authoringamap"
         minSdkVersion 16
         targetSdkVersion 23
         versionCode 1
@@ -20,8 +20,6 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    testCompile 'junit:junit:4.12'
     compile 'com.android.support:appcompat-v7:23.4.0'
     compile 'com.android.support:design:23.4.0'
 }
