+++ authoring-map/proguard-rules.pro
