+++ authoring-map/src/main/java/com/esri/arcgisruntime/sample/authoringamap/MainActivity.java
@@ -14,7 +14,7 @@
  *
  */
 
-package com.esri.arcgisruntime.sample.authoramap;
+package com.esri.arcgisruntime.sample.authoringamap;
 
 import android.app.ProgressDialog;
 import android.content.res.Configuration;
