+++ authoring-map/src/main/res/drawable-mdpi/ic_drawer.png
