+++ authoring-map/src/main/res/drawable-xhdpi/ic_drawer.png
