+++ authoring-map/src/main/res/drawable-xxhdpi/ic_drawer.png
