+++ authoring-map/src/main/res/layout/activity_drawer.xml
