+++ authoring-map/src/main/res/layout/activity_map_save.xml
