+++ authoring-map/src/main/res/menu/menu_main.xml
