+++ authoring-map/src/main/res/values/colors.xml
@@ -3,8 +3,6 @@
     <color name="primary">#4CAF50</color>
     <color name="primary_dark">#388E3C</color>
     <color name="accent">#FF5722</color>
-    <color name="text_primary">#212121</color>
-    <color name="text_secondary">#727272</color>
     <color name="list_background">#fff5f5f5</color>
     <color name="text_white">#ffffff</color>
 </resources>
