+++ authoring-map/src/main/res/values/dimens.xml
