+++ authoring-map/src/main/res/values/portal_settings.xml
@@ -4,11 +4,11 @@
     <string-array name="portal">
         <item name="title">Android Sample</item>
         <!-- your account URL path-->
-        <item name="url">https://androidteam.maps.arcgis.com/</item>
+        <item name="url">https://your-team.maps.arcgis.com/</item>
         <!-- Client ID of your registered application-->
-        <item name="client_id">BNuvI0QZuudSFeoW</item>
+        <item name="client_id">your-clientID</item>
         <!-- Redirect URI of the application from Authentication tab-->
-        <item name="redirect_uri">my-ags-app://auth</item>
+        <item name="redirect_uri">redirect-uri</item>
         <item name="username">your_user_name</item>
         <item name="password">your_password</item>
     </string-array>
