+++ authoring-map/src/main/res/values/strings.xml
