+++ authoring-map/src/main/res/values/styles.xml
