+++ settings.gradle
@@ -4,7 +4,7 @@ include ':add-graphics-renderer',
         ':arcgis-tiledlayer-rendermode',
         ':arcgis-tiledlayer-url',
         ':arcgis-vectortiledlayer-url',
-        ':author-map',
+        ':authoring-map',
         ':change-basemaps',
         ':change-feature-layer-renderer',
         ':change-sublayer-visibility',
